+++ WebUI/electron/subprocesses/apiServiceRegistry.ts
@@ -4,7 +4,7 @@ import { AiBackendService } from './aiBackendService.ts'
 import { BrowserWindow } from 'electron'
 import { appLoggerInstance } from '../logging/logger.ts'
 import getPort, { portNumbers } from 'get-port'
-import { LlamaCppBackendService } from './llamaCppBackendService.ts'
+import { LlamaCppBackendService } from './llamaCppService/llamaCppBackendService.ts'
 import { OpenVINOBackendService } from './openVINOBackendService.ts'
 import { OllamaBackendService } from './ollamaBackendService.ts'
 import { LocalSettings } from '../main.ts'
