+++ WebUI/package.json
@@ -22,10 +22,10 @@
     "@ai-sdk/vue": "^2.0.115",
     "@google/model-viewer": "^4.1.0",
     "@heroicons/vue": "^2.2.0",
-    "@langchain/classic": "^1.0.16",
     "@langchain/community": "^1.1.15",
     "@langchain/core": "^1.1.24",
     "@langchain/openai": "^1.2.6",
+    "@langchain/textsplitters": "^1.0.5",
     "@radix-icons/vue": "^1.0.0",
     "@tailwindcss/vite": "^4.1.18",
     "@vee-validate/zod": "^4.15.1",
@@ -46,7 +46,7 @@
     "marked-shiki": "^1.2.1",
     "ollama": "^0.6.3",
     "partysocket": "^1.1.6",
-    "pdf-parse": "1.1.1",
+    "pdf-parse": "^2.4.5",
     "pinia": "^3.0.4",
     "pinia-plugin-persistedstate": "^4.7.1",
     "radix-vue": "^1.9.17",
